+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.6.2 - 2026-07-14
+
+### Fixes
+
+- Map: remove tilted view setting and HUD control.
+
 ## 0.6.1 - 2026-07-14
 
 ### Improvements